Fetch BBS Occurrence Data by Species

This function fetches occurrence data for a specified target species, utilizing both the event and occurrence tables from GBIF. The fetched dataset undergoes the following processing steps compared to the raw dataset:

Retains only observations for the specified target species.

Converts implicit missing values into explicit ones (e.g., filling in zeros for trips where the target species was not observed). Specifically, if a site/plot was visited in a specific year/trip, but without observing the target species, then the dataset will have a value of 0 for that row.

Uso

bbs_fetch(target_species = NULL)

Argumentos

target_species

Character string specifying the scientific name of the species of interest. It can accept a single character string, such as target_species = "Hypsipetes leucocephalus", or a vector, such as target_species = c("Hypsipetes leucocephalus", "Heterophasia auricularis"). Leave undefined or use NULL to return all species. Use the bbs_translate function to help find the species' scientific names.

Valor

A tibble containing the species occurrence data.

Ejemplos

bbs_fetch(target_species = c("Pycnonotus taivanus", "Pycnonotus sinensis"))
#> # A tibble: 92,475 × 16
#>     year month   day site   locationID decimalLatitude decimalLongitude weather
#>    <dbl> <dbl> <dbl> <chr>  <chr>                <dbl>            <dbl> <chr>  
#>  1  2009     3    10 A02-01 A02-01_01             25.1             122. NA     
#>  2  2009     3    10 A02-01 A02-01_01             25.1             122. NA     
#>  3  2009     4     5 A02-01 A02-01_01             25.1             122. NA     
#>  4  2009     4    26 A02-01 A02-01_01             25.1             122. NA     
#>  5  2009     3    10 A02-01 A02-01_01             25.1             122. NA     
#>  6  2009     4    26 A02-01 A02-01_01             25.1             122. NA     
#>  7  2009     4     5 A02-01 A02-01_01             25.1             122. NA     
#>  8  2009     4    26 A02-01 A02-01_02             25.1             122. NA     
#>  9  2009     3    10 A02-01 A02-01_02             25.1             122. NA     
#> 10  2009     4     5 A02-01 A02-01_02             25.1             122. NA     
#> # ℹ 92,465 more rows
#> # ℹ 8 more variables: wind <chr>, habitat <chr>, scientificName <chr>,
#> #   vernacularName <chr>, individualCount <dbl>, time_slot <chr>,
#> #   distance <chr>, flock <chr>