Visualize Species Distribution Across All BBS Sites

This function visualizes the sites surveyed for breeding birds in Taiwan, highlighting the presence and absence of specific species. It is designed upon the function bbs_fetch.

Usage

bbs_plotmap(target_species)

Arguments

target_species

Character string specifying the scientific name of the species of interest. It can accept a single character string, such as target_species = "Hypsipetes leucocephalus", or a vector, such as target_species = c("Hypsipetes leucocephalus", "Heterophasia auricularis"). Leave undefined or use NULL to return no species. Use the bbs_translate function to help find the species' scientific names.

Value

A ggplot object showing the distribution map.

Examples

bbs_plotmap(target_species = c("Pycnonotus taivanus", "Pycnonotus sinensis"))
#> ! The bird is not in the BBS species list
#> ! 查無鳥名
#> ! The bird is not in the BBS species list
#> ! 查無鳥名
#> NULL