Return the Coordinates of All BBS Sites

This function returns the coordinates of all BBS sites that were surveyed. No arguments are needed. The coordinates are reported using the WGS84 projection system. Use terra::vect(geom = c("decimalLatitude", "decimalLongitude"), crs = "epsg:4326") to transform the table to spatial object.

Usage

bbs_sites()

Value

A tibble including the coordinates of all BBS survey sites, in WGS84.

Details

The source data comes from the event table in the GBIF dataset.

Examples

# Get the full list of BBS sites in a tibble
bbs_sites()
#> # A tibble: 4,160 × 6
#>    site   plot  locationID locality   decimalLatitude decimalLongitude
#>    <chr>  <chr> <chr>      <chr>                <dbl>            <dbl>
#>  1 A02-01 01    A02-01_01  台北縣瑞芳            25.1             122.
#>  2 A02-01 02    A02-01_02  台北縣瑞芳            25.1             122.
#>  3 A02-01 03    A02-01_03  台北縣瑞芳            25.1             122.
#>  4 A02-01 04    A02-01_04  台北縣瑞芳            25.1             122.
#>  5 A02-01 05    A02-01_05  台北縣瑞芳            25.1             122.
#>  6 A02-01 06    A02-01_06  台北縣瑞芳            25.1             122.
#>  7 A02-01 07    A02-01_07  台北縣瑞芳            25.1             122.
#>  8 A02-01 08    A02-01_08  台北縣瑞芳            25.1             122.
#>  9 A02-01 09    A02-01_09  台北縣瑞芳            25.1             122.
#> 10 A02-01 10    A02-01_10  台北縣瑞芳            25.1             122.
#> # ℹ 4,150 more rows

# Transform BBS sites into a spatial object using terra package
bbs_sites() |>
terra::vect(geom = c("decimalLatitude", "decimalLongitude"), crs = "epsg:4326")
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 4160, 4  (geometries, attributes)
#>  extent      : 21.94431, 25.27016, 120.0673, 121.9298  (xmin, xmax, ymin, ymax)
#>  coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#>  names       :   site  plot locationID   locality
#>  type        :  <chr> <chr>      <chr>      <chr>
#>  values      : A02-01    01  A02-01_01 台北縣瑞芳
#>                A02-01    02  A02-01_02 台北縣瑞芳
#>                A02-01    03  A02-01_03 台北縣瑞芳