Translate Bird Species' Chinese Common Name to Scientific Name

This function is intended for use under bbs_fetch and bbs_plotmap. This function helps users find the scientific names of birds from their Chinese common names for species found in Taiwan.

Usage

bbs_translate(target_species)

Arguments

target_species

A single character string or a vector of character strings representing species' names in Chinese.

Value

A vector of bird species' scientific names. If the input species name is not included in the bird list of Taiwan, NA will be returned. Please check for any typos.

Examples

# For a single species
bbs_translate("白頭翁")
#> [1] "Pycnonotus sinensis"

# For multiple species
bbs_translate(target_species = c("烏頭翁", "白頭翁", "紅嘴黑鵯", "白耳畫眉"))
#> [1] "Pycnonotus taivanus"      "Pycnonotus sinensis"     
#> [3] "Hypsipetes leucocephalus" "Heterophasia auricularis"