BBS Taiwan raw dataset on GBIF - measurementorfacts

A dataframe record BBS Taiwan measurementorfacts, which contains info for associated event record, such as time, location. This is a raw dataset, which was downloaded directly from GBIF without any wrangling

Uso

measurementorfacts

Formato

A data frame with 1,649,589 rows and 7 columns:

id

measurementID

measurementType

measurementValue

measurementDeterminedDate

measurementDeterminedBy

measurementMethod

...

Fuente

https://www.gbif.org/zh-tw/dataset/f170f056-3f8a-4ef3-ac9f-4503cc854ce0