Fetch BBS Taiwan data from GBIF

Filter and clean BBS Taiwan data given certain year range and target species

Uso

bbs_fetch(target_species = NULL)

Argumentos

target_species

a character vector for species of interest, return all species if NULL

Valor

a list containing 1. species occurrence and 2. site information

Ejemplos

bbs_fetch(target_species = c("Pycnonotus taivanus", "Pycnonotus sinensis"))
#> Joining with `by = join_by(locationID)`
#> $occurrence
#> # A tibble: 52,118 × 15
#>     year month   day locationID eventID       weather wind  habitat occurrenceID
#>    <dbl> <dbl> <dbl> <chr>      <chr>         <chr>   <chr> <chr>   <chr>       
#>  1  2009     3    10 A02-01_01  TWBBS_2009_A… NA      NA    C1:A1:… TWBBS_2009_…
#>  2  2009     3    10 A02-01_01  TWBBS_2009_A… NA      NA    C1:A1:… TWBBS_2009_…
#>  3  2009     4     5 A02-01_01  TWBBS_2009_A… NA      NA    C1:A1:… TWBBS_2009_…
#>  4  2009     4    26 A02-01_01  TWBBS_2009_A… NA      NA    C1:A1:… TWBBS_2009_…
#>  5  2009     3    10 A02-01_01  TWBBS_2009_A… NA      NA    C1:A1:… TWBBS_2009_…
#>  6  2009     4    26 A02-01_01  TWBBS_2009_A… NA      NA    C1:A1:… TWBBS_2009_…
#>  7  2009     4     5 A02-01_01  TWBBS_2009_A… NA      NA    C1:A1:… TWBBS_2009_…
#>  8  2009     4    26 A02-01_02  TWBBS_2009_A… NA      NA    C2:A1:… TWBBS_2009_…
#>  9  2009     3    10 A02-01_02  TWBBS_2009_A… NA      NA    C2:A1:… TWBBS_2009_…
#> 10  2009     4     5 A02-01_02  TWBBS_2009_A… NA      NA    C2:A1:… TWBBS_2009_…
#> # ℹ 52,108 more rows
#> # ℹ 6 more variables: scientificName <chr>, vernacularName <chr>,
#> #   individualCount <dbl>, time_slot <chr>, distance <chr>, flock <chr>
#> 
#> $site_info
#> # A tibble: 4,160 × 9
#>    site  plot  locationID locality decimalLatitude decimalLongitude  elev region
#>    <chr> <chr> <chr>      <chr>              <dbl>            <dbl> <dbl> <chr> 
#>  1 A02-… 01    A02-01_01  台北縣…             25.1             122.  109. North 
#>  2 A02-… 02    A02-01_02  台北縣…             25.1             122.  105. North 
#>  3 A02-… 03    A02-01_03  台北縣…             25.1             122.  105. North 
#>  4 A02-… 04    A02-01_04  台北縣…             25.1             122.  105. North 
#>  5 A02-… 05    A02-01_05  台北縣…             25.1             122.  105. North 
#>  6 A02-… 06    A02-01_06  台北縣…             25.1             122.  105. North 
#>  7 A02-… 07    A02-01_07  台北縣…             25.1             122.  105. North 
#>  8 A02-… 08    A02-01_08  台北縣…             25.1             122.  105. North 
#>  9 A02-… 09    A02-01_09  台北縣…             25.1             122.  154. North 
#> 10 A02-… 10    A02-01_10  台北縣…             25.1             122.  154. North 
#> # ℹ 4,150 more rows
#> # ℹ 1 more variable: zone <chr>
#>