Fetch BBS Occurrence Data by Species

This function fetches occurrence data for a specified target species, utilizing both the event and occurrence tables from GBIF. The fetched dataset undergoes the following processing steps compared to the raw dataset:

Retains only observations for the specified target species.

Converts implicit missing values into explicit ones (e.g., filling in zeros for trips where the target species was not observed). Specifically, if a site/plot was visited in a specific year/trip, but without observing the target species, then the dataset will have a value of 0 for that row.

Usage

bbs_fetch(target_species)

Arguments

target_species

Character string specifying the scientific name of the species of interest. It can accept a single character string, such as target_species = "Hypsipetes leucocephalus", or a vector, such as target_species = c("Hypsipetes leucocephalus", "Heterophasia auricularis"). Leave undefined or use NULL to return all species. Use the bbs_translate function to help find the species' scientific names.

Value

A tibble containing the species occurrence data.

Examples

bbs_fetch(target_species = c("Pycnonotus taivanus", "Pycnonotus sinensis"))
#> ! The bird is not in the BBS species list
#> ! 查無鳥名
#> ! The bird is not in the BBS species list
#> ! 查無鳥名
#> NULL